Move Qwen vision state into dedicated module
